restructure end to end tests
